fix: split response, request dto
