+++ src/main/java/com/scg/stop/domain/notice/dto/request/NoticeRequestDto.java
@@ -1,0 +1,34 @@
+package com.scg.stop.domain.notice.dto.request;
+
+import com.scg.stop.domain.notice.domain.Notice;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class NoticeRequestDto {
+
+    @NotBlank(message = "제목을 입력해주세요.")
+    private String title;
+
+    @NotBlank(message = "내용을 입력해주세요.")
+    private String content;
+
+    @NotNull(message = "고정 여부를 입력해주세요.")
+    private boolean fixed;
+
+    // DTO -> Entity
+    public Notice toEntity() {
+        Notice notice =  Notice.builder()
+                .title(title)
+                .content(content)
+                .hitCount(0)
+                .fixed(fixed)
+                .build();
+        return notice;
+    }
+}
