+++ src/main/java/com/scg/stop/domain/notice/repository/NoticeRepository.java
@@ -1,7 +1,8 @@
 package com.scg.stop.domain.notice.repository;
 
 import com.scg.stop.domain.notice.domain.Notice;
-import com.scg.stop.domain.notice.dto.NoticeDto;
+import com.scg.stop.domain.notice.dto.request.NoticeRequestDto;
+import com.scg.stop.domain.notice.dto.response.NoticeResponseDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -17,9 +18,9 @@ public interface NoticeRepository extends JpaRepository<Notice, Long> {
     @Transactional
     @Modifying
     @Query("UPDATE Notice n SET n.title = :#{#dto.title}, n.content = :#{#dto.content}, n.fixed = :#{#dto.fixed} WHERE n.id = :id")
-    int updateNotice(@Param("id") Long id, @Param("dto") NoticeDto.Request dto);
+    int updateNotice(@Param("id") Long id, @Param("dto")NoticeRequestDto dto);
 
     @Query("SELECT n FROM Notice n " +
             "WHERE (:title IS NULL OR n.title = :title)")
-    Page<NoticeDto.Response> findNotices(@Param("title") String title, Pageable pageable);
+    Page<NoticeResponseDto> findNotices(@Param("title") String title, Pageable pageable);
 }
